updates on working files

+++ demo/controller/notes/destroy.php
@@ -8,7 +8,7 @@ $db = App::resolve(Database::class);
 
 
 
-$currentuser =1;
+$currentuser=2;
 
 $note = $db->query ("SELECT * FROM notes WHERE id = :id" ,
 [
@@ -19,10 +19,13 @@ $note = $db->query ("SELECT * FROM notes WHERE id = :id" ,
 
  authorize($note['user_id']==$currentuser);
 
-    
+   if($currentuser=='user_id'){
     $db->query(" DELETE FROM notes WHERE id =:id", [
         'id' => $_POST['id']
     ]);
+
+   } 
+    
 
 
     header('location: /notes');
